Added FuncInfo3 exception support
This parses the FuncInfo3 exception info so that we can support it

It is very similar to the FH4 format just without the compressed encodings.

Renamed pe_ip2state_entry_t to pe_fh4_encoded_entry as it is more accurate naming and specifies that it pertains only to FH4

Also improved update_rva_in_assembler_instruction by returning false if no operand was updated.

+++ ext/binwrite/binary/pe/pe_frame_pointer.cpp
@@ -345,8 +345,8 @@ static void apply_frame_pointer_unwind_info(binwrite::portable_executable_t& pe,
 				continue;
 			}
 
-			const std::uint32_t codes_end_offset = codes_end - reinterpret_cast<const std::uint8_t*>(unwind_info);
-			const std::uint32_t codes_start_offset = codes_start - reinterpret_cast<const std::uint8_t*>(unwind_info);
+			const std::uint32_t codes_end_offset = static_cast<std::uint32_t>(codes_end - reinterpret_cast<const std::uint8_t*>(unwind_info));
+			const std::uint32_t codes_start_offset = static_cast<std::uint32_t>(codes_start - reinterpret_cast<const std::uint8_t*>(unwind_info));
 
 			copied_unwind_info.insert_range(copied_unwind_info.begin() + codes_end_offset, end_bytes);
 			copied_unwind_info.insert_range(copied_unwind_info.begin() + codes_start_offset, start_bytes);
